% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DVS-package.R
\docType{package}
\name{DVS-package}
\alias{DVS-package}
\title{DVS: Decorrelation for Variable Selection}
\description{
Tools for decorrelation-based variable selection using Air-HOLP,
Gram-Schmidt orthogonalization, and stability selection with the Lasso.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MahdiNouraie/DVS}
  \item \doi{10.1007/s11222-026-10916-7}
  \item Report bugs at \url{https://github.com/MahdiNouraie/DVS/issues}
}

}
\author{
\strong{Maintainer}: Mahdi Nouraie \email{mahdinouraie20@gmail.com}

Authors:
\itemize{
  \item Connor Smith
  \item Samuel Muller
}

}
