% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{dvs_init}
\alias{dvs_init}
\title{initialize dvs to add/get versioned files from the storage directory}
\usage{
dvs_init(storage_directory, permissions = NULL, group = NULL)
}
\arguments{
\item{storage_directory}{path to the desired storage directory for versioned files}

\item{permissions}{optional: linux file permissions to set
for files added to the storage directory (in octal format)}

\item{group}{optional: primary group to set for files added to the
storage directory}
}
\value{
A data frame with the storage directory, permissions, and group
}
\description{
initialize dvs to add/get versioned files from the storage directory
}
\examples{
\dontrun{
# would initialize the project's storage directory at /data/project_x
dvs_init("/data/project_x")

# would initialize the project's storage directory at /data/project_x and
# configure the linux permissions "777" and primary group "project_x_group" for all
# files added to the storage directory
dvs_init("/data/project_x", 777, "project_x_group")
}

}
