% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{dvs-package}
\alias{dvs}
\alias{dvs-package}
\title{dvs: Large/Sensitive File Linker for Git Repositories}
\description{
This package allows users to version large or sensitive files under Git without tracking the file directly.
}
\author{
\strong{Maintainer}: Jenna Johnson \email{jenna@a2-ai.com}

Authors:
\itemize{
  \item Devin Pastoor \email{devin@a2-ai.com}
}

}
