% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_functions.R
\name{dvs_add}
\alias{dvs_add}
\title{add files to the storage directory}
\usage{
dvs_add(files, message = "")
}
\arguments{
\item{files}{files to add to the storage directory}

\item{message}{optional argument to add a message to future data frame rows associated with these files}
}
\value{
a data frame with the states of success of added files
}
\description{
add files to the storage directory
}
