% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_functions.R
\name{dvs_get}
\alias{dvs_get}
\title{get files previously added to the storage directory}
\usage{
dvs_get(files)
}
\arguments{
\item{files}{files to get from the storage directory}
}
\value{
a data frame with the states of success of retrieved files
}
\description{
get files previously added to the storage directory
}
