% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_functions.R
\name{dvs_get}
\alias{dvs_get}
\title{get added files}
\usage{
dvs_get(files)
}
\arguments{
\item{files}{file paths or glob patterns to get from the storage directory}
}
\value{
a data frame with the states of success of retrieved files
}
\description{
get added files
}
\details{
retrieves files previously added with \link{dvs_add} to the storage directory (initialized by \link{dvs_init}).
}
\examples{
\dontrun{
# would get all previously added files in data/derived from
# the initialized storage directory
dvs_get("data/derived/*.dvsmeta")

# would get all files in data/derived (excluding
# .gitignore files) and ~Projects/project_x/large_file.pdf
# from the initialized storage directory
dvs_get(c("data/derived/*", "~Projects/project_x/large_file.pdf"))
}

}
