% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_functions.R
\name{dvs_init}
\alias{dvs_init}
\title{initialize devious}
\usage{
dvs_init(storage_directory, permissions = 664, group = "")
}
\arguments{
\item{storage_directory}{path to storage directory}

\item{permissions}{optional argument to set linux file permissions (in octal format)}

\item{group}{optional argument to set group of files stored}
}
\value{
nothing, or errors initializing
}
\description{
initialize devious
}
