% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_functions.R
\name{dvs_init}
\alias{dvs_init}
\title{initialize devious}
\usage{
dvs_init(directory, permissions = 511, group = "")
}
\arguments{
\item{directory}{path to storage directory}

\item{permissions}{optional argument to set permissions of files stored}

\item{group}{optional argument to set group of files stored}
}
\value{
nothing, or errors initializing
}
\description{
initialize devious
}
