% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_functions.R
\name{dvs_status}
\alias{dvs_status}
\title{get the status of any file previously added to the storage directory}
\usage{
dvs_status(files = c())
}
\arguments{
\item{files}{optional argument: when specified, returns data frame with only these specified file paths or glob patterns}
}
\value{
a data frame with the statuses of all previously added files
}
\description{
get the status of any file previously added to the storage directory
}
