% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_functions.R
\name{dvs_status}
\alias{dvs_status}
\title{status report for added files}
\usage{
dvs_status(files = c(""), split_output = FALSE)
}
\arguments{
\item{files}{optional: when specified, returns data frame with only these specified file paths or glob patterns.}

\item{split_output}{when split_output is true, a list of two data frames - \code{successes} and \code{failures} - is returned.
Rows in \code{successes} are files with successfully curated statuses, and rows in \code{failures} are inputs that returned errors.
When split_output is false, the output is a single data frame with all attempted file status reports,
and whose success or failure is indicated as such in the outcome column.}
}
\value{
a data frame with the statuses of previously added files
}
\description{
status report for added files
}
\details{
gives the statuses of previously added files (\code{up-to-date}, \code{out-of-sync}, or \code{not-present})
to make users aware if files stored in the storage directory don't exist in their local repository or have been updated.
If no file paths or glob patterns are inputted, \code{dvs_status} gives the status of all previously added files.
If there an error getting the status of a particular file, the function itself will not return an error, rather, a given error will be indicated in the data frame output.
}
\examples{
\dontrun{
  # would give the status of all previously added files
  dvs_status()

  # would attempt to get the status of all files in data/derived (except for .gitignore files)
  dvs_status("data/derived/*")
}

}
